+++ app/src/main/java/com/jHerscu/clearskies/ClearSkiesApp.kt
@@ -1,0 +1,16 @@
+package com.jHerscu.clearskies
+
+import android.app.Application
+import com.jHerscu.clearskies.BuildConfig
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class ClearSkiesApp : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
+}
