--- app/src/main/java/com/jHerscu/clearskies/di/ClearSkiesApp.kt
@@ -1,7 +1,0 @@
-package com.jHerscu.clearskies.di
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class ClearSkiesApp : Application()
