--- app/src/main/java/com/jHerscu/clearskies/di/ViewModelModule.kt
@@ -1,10 +1,0 @@
-package com.jHerscu.clearskies.di
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-object ViewModelModule {
-}
